+++ src/ExAs/Assertions/MemberAssertions/Numbers/LessOrEqualAssertion.cs
@@ -6,12 +6,12 @@ using ExAs.Utils.SystemExtensions;
 
 namespace ExAs.Assertions.MemberAssertions.Numbers
 {
-    public class GreaterOrEqualAssertion<T> : IAssertValue<T>
+    public class LessOrEqualAssertion<T> : IAssertValue<T>
         where T : IComparable<T>
     {
         private readonly T expected;
 
-        public GreaterOrEqualAssertion(T expected)
+        public LessOrEqualAssertion(T expected)
         {
             this.expected = expected;
         }
@@ -19,9 +19,9 @@ namespace ExAs.Assertions.MemberAssertions.Numbers
         public ValueAssertionResult AssertValue(T actual)
         {
             return new ValueAssertionResult(
-                actual.Compare_NullAware(expected) >= 0, 
+                actual.Compare_NullAware(expected) <= 0,
                 actual.ToValueString(),
-                ComposeLog.Expected($"greater or equal to {expected.ToValueString()}"));
+                ComposeLog.Expected($"less or equal to {expected.ToValueString()}"));
         }
     }
 }